Remove the guts of loop()
This patch removes the majority of loop(), in favour of two things.

First, we echo any important sensor state (like gravity, any IR
commands, and battery voltage level) to the serial line once a tick,
with each tick's frame separated with `"-- \n"`.

Second, we introduce a teeny little interpreter so we can read commands
over the serial line in a reasonable syntax.  That syntax is Forth.
Right now it's limited in that you can't use ':' so you can't define
your own words; that'll come later.  For now it's a convenient way to
register native commands.

The idea is that by echoing out the state, receiving commands over the
serial line, and running loop() as close as possible to once every 50ms,
we can move complex processing off to the piggy-back raspberry pi.  Part
of the state we echo out is the start time of the loop iteration in
milliseconds, so we can detect whether the loop timing is being
violated.