Move loop control to the host
This patch does away with the loop() loop, in favour of relying on the
host to send commands and control the timing adequately, presenting a
'> ' prompt.

This allows the host to send lines as fast as the nybble will
successfully respond, without assuming that the loop needs to do a full
delay cycle before being able to process the next command.